pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-toolkit"
version = "0.1.0"
description = "Bridge LLM workflows to the traditional ways of working"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Ahmed Khaled Mohamed", email = "ahmd.khaled.a.mohamed@gmail.com"},
]
keywords = ["llm", "markdown", "google-docs", "productivity", "ai-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "markdown",
    "google-auth-oauthlib",
    "google-api-python-client",
]

[project.urls]
Homepage = "https://github.com/ahmedkhaledmohamed/llm-toolkit"
Issues = "https://github.com/ahmedkhaledmohamed/llm-toolkit/issues"

[project.scripts]
llmtk = "llmtk.cli:main"

[tool.setuptools.packages.find]
include = ["llmtk*"]
